refactor(website): extract landing into its own module
