+++ website/src/landing/hero.rs
@@ -1,0 +1,6 @@
+use dioxus::prelude::*;
+
+#[allow(dead_code)]
+pub fn placeholder() -> Element {
+    rsx! {}
+}
