+++ website/src/landing/platform.rs
@@ -1,0 +1,6 @@
+use dioxus::prelude::*;
+
+#[allow(dead_code)]
+pub fn placeholder() -> Element {
+    rsx! {}
+}
